+++ src/main/java/de/samply/fhir/BundleToContainersConverter.java
@@ -2,17 +2,21 @@ package de.samply.fhir;
 
 
 import ca.uhn.fhir.context.FhirContext;
-import de.samply.converter.ConverterImpl;
 import de.samply.container.Attribute;
 import de.samply.container.Container;
 import de.samply.container.Containers;
+import de.samply.converter.ConverterImpl;
 import de.samply.converter.Format;
 import de.samply.template.AttributeTemplate;
 import de.samply.template.ContainerTemplate;
 import de.samply.template.ConverterTemplate;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
 import org.hl7.fhir.r4.hapi.ctx.HapiWorkerContext;
+import org.hl7.fhir.r4.model.Base;
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.ExpressionNode;
 import org.hl7.fhir.r4.model.Resource;
@@ -36,23 +40,42 @@ public class BundleToContainersConverter extends ConverterImpl<Bundle, Container
 
   public Containers convertToContainers(Bundle bundle, ConverterTemplate converterTemplate) {
     Containers containers = new Containers();
+    Map<String, Resource> idResourceMap = fetchIdResourceMap(bundle);
     if (converterTemplate != null) {
       converterTemplate.getContainerTemplates()
-          .forEach(containerTemplate -> addContainers(bundle, containers, containerTemplate));
+          .forEach(containerTemplate -> addContainers(bundle, containers, containerTemplate,
+              idResourceMap));
     }
     return containers;
   }
 
+  private Map<String, Resource> fetchIdResourceMap(Bundle bundle) {
+    Map<String, Resource> result = new HashMap<>();
+    bundle.getEntry().forEach(bundleEntryComponent -> {
+      Resource resource = bundleEntryComponent.getResource();
+      String id = resource.getResourceType() + "/" + resource.getIdPart();
+      result.put(id, resource);
+    });
+    return result;
+  }
+
   private void addContainers(Bundle bundle, Containers containers,
-      ContainerTemplate containerTemplate) {
+      ContainerTemplate containerTemplate, Map<String, Resource> idResourceMap) {
     containerTemplate.getAttributeTemplates().forEach(attributeTemplate ->
         bundle.getEntry().forEach(bundleEntryComponent -> {
-          if (isSameResourceType(bundleEntryComponent.getResource(), attributeTemplate.getFhirPath())) {
+          if (isSameResourceType(bundleEntryComponent.getResource(), attributeTemplate)) {
             fetchResourceAttribute(bundleEntryComponent.getResource(), containerTemplate,
-                attributeTemplate).forEach(resourceAttribute ->
+                attributeTemplate, idResourceMap).forEach(resourceAttribute ->
                 addResourceAttributeToContainers(containers, resourceAttribute));
           }
         }));
+  }
+
+  private boolean isSameResourceType(Resource resource, AttributeTemplate attributeTemplate) {
+    return (attributeTemplate.getParentFhirPath() != null &&
+        isSameResourceType(resource, attributeTemplate.getParentFhirPath())) ||
+        isSameResourceType(resource, attributeTemplate.getFhirPath());
+
   }
 
   private boolean isSameResourceType(Resource resource, String fhirPath) {
@@ -67,25 +90,56 @@ public class BundleToContainersConverter extends ConverterImpl<Bundle, Container
   }
 
   private List<ResourceAttribute> fetchResourceAttribute(Resource resource,
-      ContainerTemplate containerTemplate, AttributeTemplate attributeTemplate) {
+      ContainerTemplate containerTemplate, AttributeTemplate attributeTemplate,
+      Map<String, Resource> idResourceMap) {
     List<ResourceAttribute> resourceAttributes = new ArrayList<>();
-    ExpressionNode expressionNode = fhirPathEngine.parse(attributeTemplate.getFhirPath());
-    fhirPathEngine.evaluate(resource, expressionNode)
-        .forEach(base -> resourceAttributes.add(
-            new ResourceAttribute(resource, base.toString(), containerTemplate,
-                attributeTemplate)));
+    Resource realResource = fetchRealResource(resource, attributeTemplate, idResourceMap);
+    if (realResource != null) {
+      ExpressionNode expressionNode = fhirPathEngine.parse(attributeTemplate.getFhirPath());
+      fhirPathEngine.evaluate(realResource, expressionNode)
+          .forEach(base -> resourceAttributes.add(
+              new ResourceAttribute(resource, base.toString(), containerTemplate,
+                  attributeTemplate)));
+    }
     return resourceAttributes;
+  }
+
+  private Resource fetchRealResource(Resource currentResource, AttributeTemplate attributeTemplate,
+      Map<String, Resource> idResourceMap) {
+    AtomicReference<Resource> result = new AtomicReference<>(currentResource);
+    if (attributeTemplate.getParentFhirPath() != null) {
+      attributeTemplate.fetchParentFhirPaths().forEach(parentFhirPath -> {
+        if (result.get() != null) {
+          result.set(fetchParentResource(result.get(), parentFhirPath, idResourceMap));
+        }
+      });
+    }
+    return result.get();
+  }
+
+  private Resource fetchParentResource(Resource resource, String parentFhirPath,
+      Map<String, Resource> idResourceMap) {
+    Resource result = null;
+    ExpressionNode expressionNode = fhirPathEngine.parse(parentFhirPath);
+    List<Base> resourceIds = fhirPathEngine.evaluate(resource, expressionNode);
+    if (resourceIds.size() > 0) {
+      String resourceId = resourceIds.get(0).toString();
+      result = idResourceMap.get(resourceId);
+    }
+    return result;
   }
 
   private void addResourceAttributeToContainers(Containers containers,
       ResourceAttribute resourceAttribute) {
     Container container = containers.getContainer(resourceAttribute.containerTemplate(),
         resourceAttribute.fetchContainerId());
-    if (container == null){
-      container = new Container(resourceAttribute.fetchContainerId(), resourceAttribute.containerTemplate());
+    if (container == null) {
+      container = new Container(resourceAttribute.fetchContainerId(),
+          resourceAttribute.containerTemplate());
       containers.addContainer(container);
     }
-    container.addAttribute(new Attribute(resourceAttribute.attributeTemplate(), resourceAttribute.attributeValue()));
+    container.addAttribute(
+        new Attribute(resourceAttribute.attributeTemplate(), resourceAttribute.attributeValue()));
   }
 
   private FHIRPathEngine createFhirPathEngine() {